bottomNavigationBarを追加
bottomNavigationBarを追加
ファイルを分割
スタンプページのガワを作成
