+++ lib/main.dart
@@ -1,8 +1,8 @@
-import 'package:egp_client/view/shop_detail.dart';
+import 'package:egp_client/view/shop_detail_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'view/home.dart';
+import 'view/home_page.dart';
 
 void main() async {
   await dotenv.load(fileName: '.env');
@@ -19,8 +19,8 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Flutter Google Maps Sample',
       routes: <String, WidgetBuilder>{
-        '/': (_) => Home(),
-        '/shop_detail': (_) => ShopDetail(
+        '/': (_) => HomePage(),
+        '/shop_detail': (_) => ShopDetailPage(
               year: 0,
               no: 0,
               shopName: '',
